+++ backend/src/main/java/com/java/NBE4_5_1_8/domain/item/repository/ItemRepository.java
@@ -1,6 +1,6 @@
 package com.java.NBE4_5_1_8.domain.item.repository;
 
-import com.java.NBE4_5_1_8.domain.item.entiry.Item;
+import com.java.NBE4_5_1_8.domain.item.entity.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ItemRepository extends JpaRepository<Item, Long> {
