+++ backend/src/main/java/com/java/NBE4_5_1_8/domain/orderInfo/entity/OrderInfo.java
@@ -1,4 +1,4 @@
-package com.java.NBE4_5_1_8.domain.item.entiry;
+package com.java.NBE4_5_1_8.domain.orderInfo.entity;
 
 import com.java.NBE4_5_1_8.global.entity.BaseTime;
 import jakarta.persistence.Entity;
@@ -8,15 +8,16 @@ import jakarta.persistence.Id;
 import lombok.Getter;
 import lombok.Setter;
 
+
 @Entity
 @Getter
 @Setter
-public class Item extends BaseTime {
+public class OrderInfo extends BaseTime {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String itemName;
-    private String category;
-    private String description; // 제품 설명
-    private int stockQuantity;
+    private Long orderId;
+    private String orderStatus;
+    private String memberEmail;
+    private String memberAddress;
 }
